+++ conda/platform_darwin_amd64.go
@@ -44,7 +44,7 @@ func CondaPaths(prefix string) []string {
 }
 
 func MicromambaLink() string {
-	return settings.Global.DownloadsLink("micromamba/v0.25.1/macos64/micromamba")
+	return settings.Global.DownloadsLink(micromambaLink("macos64", "micromamba"))
 }
 
 func IsWindows() bool {